requirements.txt
uvicorn
fastapi
python-dotenv
python-multipart
aiofiles
jinja2